feat: Complete SSML/audio pipeline refactor with centralized services
- Refactor Supabase Edge workers to use centralized SSMLService/TTSService
- Add ssml_prompt persistence to analysis_ssml_segments table
- Honor AI_ANALYSIS_MODE environment variable in workers
- Update storage paths to use unified audio format resolution
- Change video_recordings.duration_seconds to numeric for decimal precision
- Update Gemini services with enhanced error handling and logging
- Move completed SSML/audio action plan to completed folder
- Update TRD and status documentation with pipeline changes

+++ packages/api/src/validation/cameraRecordingSchemas.ts
@@ -16,7 +16,7 @@ export const VideoRecordingSchema = z.object({
   filename: z.string().min(1).max(255),
   original_filename: z.string().min(1).max(255).nullable(),
   file_size: z.number().int().positive(),
-  duration_seconds: z.number().int().min(1).max(60),
+  duration_seconds: z.number().min(1).max(60),
   format: VideoFormatSchema,
   storage_path: z.string().min(1),
   upload_status: UploadStatusSchema,
